polish(copilotkit): code standardization, design polish, component modularization
