+++ src/components/ck/empty-chat-disclaimer.tsx
@@ -1,0 +1,8 @@
+/** Suppresses CopilotKit's default input disclaimer in showcase examples. */
+export function EmptyChatDisclaimer() {
+  return null;
+}
+
+export const chatInputWithoutDisclaimer = {
+  disclaimer: () => <div className="h-10" />,
+};
