+++ src/index.css
@@ -132,13 +132,3 @@
 .list-inside {
   list-style-position: inside;
 }
-
-.copilotKitMessagesContainer {
-  padding: 0 !important;
-}
-
-.copilotKitMessagesFooter {
-  width: 100% !important;
-  padding: 0.75rem 0 !important;
-  margin: 8px auto 0 auto;
-}
